refactor: move CanvasData to its own file
