+++ lib/logic/canvas_service.dart
@@ -1,3 +1,4 @@
+import 'package:editicert/models/canvas_data.dart';
 import 'package:flutter/material.dart';
 
 class CanvasService {
@@ -19,34 +20,6 @@ class CanvasService {
       hidden: backgroundHidden,
       opacity: backgroundOpacity,
       size: backgroundSize,
-    );
-  }
-}
-
-class CanvasData {
-  CanvasData({
-    required this.color,
-    required this.hidden,
-    required this.opacity,
-    required this.size,
-  });
-
-  Color color;
-  bool hidden;
-  double opacity;
-  Size size;
-
-  CanvasData copyWith({
-    Color? color,
-    bool? hidden,
-    double? opacity,
-    Size? size,
-  }) {
-    return CanvasData(
-      color: color ?? this.color,
-      hidden: hidden ?? this.hidden,
-      opacity: opacity ?? this.opacity,
-      size: size ?? this.size,
     );
   }
 }
